address comment by offline discussion
