+++ pkg/apiutil/apiutil.go
@@ -15,6 +15,7 @@
 package apiutil
 
 import (
+	"bytes"
 	"encoding/json"
 	"fmt"
 	"io"
@@ -199,4 +200,42 @@ func GetRouteName(req *http.Request) string {
 		return route.GetName()
 	}
 	return ""
+}
+
+// PostJSON is used to send POST requst to specific url
+func PostJSON(client *http.Client, url string, data []byte, checkOpts ...func([]byte, int)) error {
+	req, err := http.NewRequest(http.MethodPost, url, bytes.NewBuffer(data))
+	if err != nil {
+		return err
+	}
+	req.Header.Set("Content-Type", "application/json")
+	return doJSON(client, req, checkOpts...)
+}
+
+// GetJSON is used to send GET requst to specific url
+func GetJSON(client *http.Client, url string, data []byte, checkOpts ...func([]byte, int)) error {
+	req, err := http.NewRequest(http.MethodGet, url, bytes.NewBuffer(data))
+	if err != nil {
+		return err
+	}
+	return doJSON(client, req, checkOpts...)
+}
+
+func doJSON(client *http.Client, req *http.Request, checkOpts ...func([]byte, int)) error {
+	resp, err := client.Do(req)
+	if err != nil {
+		return errors.WithStack(err)
+	}
+	defer resp.Body.Close()
+	res, err := io.ReadAll(resp.Body)
+	if err != nil {
+		return err
+	}
+	if resp.StatusCode != http.StatusOK {
+		return errors.New(string(res))
+	}
+	for _, opt := range checkOpts {
+		opt(res, resp.StatusCode)
+	}
+	return nil
 }
