+++ tools.go
@@ -18,6 +18,8 @@
 package tools
 
 import (
+	_ "github.com/AlekSi/gocov-xml"
+	_ "github.com/axw/gocov/gocov"
 	_ "github.com/mgechev/revive"
 	_ "github.com/pingcap/errors/errdoc-gen"
 	_ "github.com/pingcap/failpoint/failpoint-ctl"
@@ -25,6 +27,4 @@ import (
 	_ "github.com/swaggo/swag/cmd/swag"
 	_ "golang.org/x/tools/cmd/goimports"
 	_ "gotest.tools/gotestsum"
-	_ "github.com/AlekSi/gocov-xml"
-	_ "github.com/axw/gocov/gocov"
 )
